Subir, listar y eliminar documentos

+++ src/app/view-user/view-user.component.ts
@@ -2,6 +2,10 @@ import { Component, OnInit } from '@angular/core';
 import { ApiService } from '../api.service';
 import {Router} from '@angular/router';
 import { Users } from '../users';
+import { faTrashAlt } from '@fortawesome/free-solid-svg-icons';
+import { faEdit } from '@fortawesome/free-solid-svg-icons';
+import { faPlusCircle } from '@fortawesome/free-solid-svg-icons';
+import { faUserShield } from '@fortawesome/free-solid-svg-icons';
 
 
 
@@ -11,9 +15,13 @@ import { Users } from '../users';
   styleUrls: ['./view-user.component.css']
 })
 export class ViewUserComponent implements OnInit {
-
+  faDelete = faTrashAlt;
+  faEdit = faEdit; 
+  faAdd = faPlusCircle;
+  faAdmin = faUserShield;
   users: any;
   tipo: any;
+  email: any;
 
   constructor(private apiService : ApiService,
               private router : Router) {
@@ -34,10 +42,16 @@ export class ViewUserComponent implements OnInit {
 
 
   deleteUser(user : Users): void {
-    this.apiService.deleteUser(user.id)
-    .subscribe( data => {
-      this.users = this.users.filter((u: Users)  => u !== user);
-    });
+    this.email = this.apiService.getEmail();
+    if(this.email == user.email){
+      alert("No te puedes eliminar tu mismo");
+    }
+    else{
+      this.apiService.deleteUser(user.id)
+      .subscribe( data => {
+        this.users = this.users.filter((u: Users)  => u !== user);
+      });
+    }
   }
 
   editUser(user: Users): void {
